TDs/8/requirements.txt
astroid==2.3.3
functional==0.4
isort==4.3.21
joblib==0.14.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
Pillow==6.2.1
pylint==2.4.4
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn==0.0
torch==1.3.1
torchvision==0.4.2
typed-ast==1.4.0
wrapt==1.11.2
